Revert 3 patches from the Linux kernel that cause a change in behavior for bound sockets
3 patches that went in for the 5.0 kernel changed the behavior of how a
socket gets picked when a packet is received, if the sockets are not in
a VRF and if one socket is bound to an interface and the other socket
isn't. The primary commit that started this issue is
torvalds/linux@6da5b0f027a825df2aebc1927a27bda185dc03d4 which prefers an
unbound datagram socket over a bound datagram socket to be chosen if the
sockets aren't in a VRF.

This breaks some assumptions that dhcrelay (from isc-dhcp) makes when it
opens a bound socket for each interface and an unbound "fallback"
socket, and it expects that the bound sockets will be preferred over the
unbound sockets.

Therefore, revert that commit (and two dependent commits that went in
after that) to restore the old behavior.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>