arm64-dts/marvell: cn9131-db-comexpress with regulator-fixed
The legacy cn9131-db-comexpress.dts configures "regulator-gpio".
This legacy is defined min/max=1.8V and works on Kernel-6.1 jsut
because the power 1.8V already set.
In fact, the regulator is Not present on board, the TRXIES Kernel-6.12
gpio probe waits for regulator, never passes and MMC/SDHCI is never activated.

FIX:
Declare "regulator-fixed" instead of "regulator-gpio".

Signed-off-by: Yan Markman <ymarkman@marvell.com>